refactor: DTO에 검증 어노테이션 추가 및 GlobalExceptionHandler 패키지 변경
