+++ src/main/java/roomescape/exception/GlobalExceptionHandler.java
@@ -1,9 +1,8 @@
-package roomescape.controller;
+package roomescape.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import roomescape.exception.NotFoundException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
