+++ src/main/java/roomescape/request/ReservationTimeRequest.java
@@ -1,10 +1,11 @@
 package roomescape.request;
 
+import jakarta.validation.constraints.NotNull;
 import roomescape.domain.ReservationTime;
 
 import java.time.LocalTime;
 
-public record ReservationTimeRequest(LocalTime startAt) {
+public record ReservationTimeRequest(@NotNull LocalTime startAt) {
     public ReservationTime toDomainForSave() {
         return new ReservationTime(null, startAt);
     }
