+++ src/main/java/roomescape/request/ThemeRequest.java
@@ -1,8 +1,13 @@
 package roomescape.request;
 
+import jakarta.validation.constraints.NotBlank;
 import roomescape.domain.Theme;
 
-public record ThemeRequest(String name, String description, String thumbnailUrl) {
+public record ThemeRequest(
+        @NotBlank String name,
+        @NotBlank String description,
+        @NotBlank String thumbnailUrl) {
+
     public Theme toDomainForSave() {
         return new Theme(null, name, description, thumbnailUrl);
     }
